+++ lib/src/widgets/root/loading/fading.dart
@@ -55,9 +55,7 @@ class _SpinKitFadingFourState extends State<SpinKitFadingFour>
       child: SizedBox.fromSize(
           size: Size.square(widget.size),
           child: Stack(
-
-              ///  ignore: always_specify_types
-              children: List.generate(4, (int i) {
+              children: List<Widget>.generate(4, (int i) {
             final double _position = widget.size * .5;
             return Positioned.fill(
                 left: _position,